feat: add global resource banner to all pages (#3046)
Move the resource banner (FAQ, Textes de référence, Nous contacter) from
the aide page into the root layout so it appears on every page. Replace
the duplicated HomeResources and AideBottomBanner components with a
single shared ResourceBanner in modules/layout.
