--- packages/app/src/modules/aide/AideBottomBanner.tsx
@@ -1,82 +1,0 @@
-import Image from "next/image";
-import Link from "next/link";
-
-import { DsfrPictogram } from "~/modules/home";
-
-type ResourceTileProps = {
-	href: string;
-	pictogramPath: string;
-	title: string;
-	detail: string;
-};
-
-function ResourceTile({
-	href,
-	pictogramPath,
-	title,
-	detail,
-}: ResourceTileProps) {
-	return (
-		<div className="fr-tile fr-tile--horizontal fr-tile--sm fr-enlarge-link">
-			<div className="fr-tile__body">
-				<div className="fr-tile__content">
-					<h3 className="fr-tile__title">
-						<Link href={href}>{title}</Link>
-					</h3>
-					<p className="fr-tile__detail">{detail}</p>
-				</div>
-			</div>
-			<div className="fr-tile__header">
-				<div className="fr-tile__pictogram">
-					<DsfrPictogram path={pictogramPath} />
-				</div>
-			</div>
-		</div>
-	);
-}
-
-/** Bottom navigation banner with links to FAQ, reference texts and contact. */
-export function AideBottomBanner() {
-	return (
-		<>
-			<div className="fr-grid-row fr-grid-row--gutters fr-mt-4w">
-				<div className="fr-col-12 fr-col-md-4">
-					<ResourceTile
-						detail="Réponses aux questions les plus courantes"
-						href="/faq"
-						pictogramPath="/dsfr/artwork/pictograms/system/information.svg"
-						title="Questions fréquentes (FAQ)"
-					/>
-				</div>
-				<div className="fr-col-12 fr-col-md-4">
-					<ResourceTile
-						detail="Consultez les textes législatifs et réglementaires"
-						href="/textes-reference"
-						pictogramPath="/dsfr/artwork/pictograms/document/document.svg"
-						title="Textes de référence"
-					/>
-				</div>
-				<div className="fr-col-12 fr-col-md-4">
-					<ResourceTile
-						detail="Besoin d'aide ? Contactez nos services d'assistance"
-						href="/aide/nous-contacter"
-						pictogramPath="/dsfr/artwork/pictograms/digital/avatar.svg"
-						title="Nous contacter"
-					/>
-				</div>
-			</div>
-			<div
-				aria-hidden="true"
-				className="fr-grid-row fr-grid-row--center fr-mt-4w"
-			>
-				<Image
-					alt=""
-					height={147}
-					src="/assets/images/home/help-illustration.svg"
-					unoptimized
-					width={210}
-				/>
-			</div>
-		</>
-	);
-}
