+++ packages/app/src/modules/home/HomePage.tsx
@@ -2,7 +2,6 @@ import { HomeHero } from "./HomeHero";
 import { HomeNotice } from "./HomeNotice";
 import styles from "./HomePage.module.scss";
 import { HomePlaceholder } from "./HomePlaceholder";
-import { HomeResources } from "./HomeResources";
 import { HomeSearch } from "./HomeSearch";
 
 /** Home page — visual content only, without tRPC provider. */
@@ -19,7 +18,6 @@ export function HomePage() {
 				</div>
 			</section>
 			<HomeSearch />
-			<HomeResources />
 		</main>
 	);
 }
