--- packages/app/src/modules/home/HomeResources.module.scss
@@ -1,14 +1,0 @@
-.section {
-	background: var(--background-alt-blue-france);
-}
-
-.illustrationWrapper {
-	display: flex;
-	justify-content: center;
-}
-
-.illustration {
-	height: auto;
-	max-width: 100%;
-	width: auto;
-}
