--- packages/app/src/modules/home/HomeResources.tsx
@@ -1,92 +1,0 @@
-import Image from "next/image";
-
-import { DsfrPictogram } from "./DsfrPictogram";
-import styles from "./HomeResources.module.scss";
-
-type ResourceTileProps = {
-	href: string;
-	pictogramPath: string;
-	title: string;
-	detail: string;
-};
-
-function ResourceTile({
-	href,
-	pictogramPath,
-	title,
-	detail,
-}: ResourceTileProps) {
-	return (
-		<div className="fr-tile fr-tile--horizontal fr-tile--sm fr-enlarge-link">
-			<div className="fr-tile__body">
-				<div className="fr-tile__content">
-					<h3 className="fr-tile__title">
-						<a href={href}>{title}</a>
-					</h3>
-					<p className="fr-tile__detail">{detail}</p>
-				</div>
-			</div>
-			<div className="fr-tile__header">
-				<div className="fr-tile__pictogram">
-					<DsfrPictogram path={pictogramPath} />
-				</div>
-			</div>
-		</div>
-	);
-}
-
-/** Resources section: FAQ, reference texts and contact. */
-export function HomeResources() {
-	return (
-		<section aria-label="Ressources utiles" className={styles.section}>
-			<div className="fr-container fr-py-6w">
-				<div className="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
-					<div className="fr-col-12 fr-col-md-9">
-						<div className="fr-grid-row fr-grid-row--gutters">
-							<div className="fr-col-12 fr-col-md-4">
-								<ResourceTile
-									detail="Réponses aux questions les plus courantes"
-									href="/faq"
-									pictogramPath="/dsfr/artwork/pictograms/system/information.svg"
-									title="Questions fréquentes (FAQ)"
-								/>
-							</div>
-
-							<div className="fr-col-12 fr-col-md-4">
-								<ResourceTile
-									detail="Consultez les textes législatifs et réglementaires"
-									href="/textes-reference"
-									pictogramPath="/dsfr/artwork/pictograms/document/document.svg"
-									title="Textes de référence"
-								/>
-							</div>
-
-							<div className="fr-col-12 fr-col-md-4">
-								<ResourceTile
-									detail="Besoin d&apos;aide ? Contactez nos services d&apos;assistance"
-									href="/contact"
-									pictogramPath="/dsfr/artwork/pictograms/digital/avatar.svg"
-									title="Nous contacter"
-								/>
-							</div>
-						</div>
-					</div>
-
-					<div
-						aria-hidden="true"
-						className={`fr-col-12 fr-col-md-3 ${styles.illustrationWrapper}`}
-					>
-						<Image
-							alt=""
-							className={styles.illustration}
-							height={147}
-							src="/assets/images/home/help-illustration.svg"
-							unoptimized
-							width={210}
-						/>
-					</div>
-				</div>
-			</div>
-		</section>
-	);
-}
