+++ packages/app/src/modules/home/index.ts
@@ -3,5 +3,5 @@ export { HomeHero } from "./HomeHero";
 export { HomeNotice } from "./HomeNotice";
 export { HomePage } from "./HomePage";
 export { HomePlaceholder } from "./HomePlaceholder";
-export { HomeResources } from "./HomeResources";
+
 export { HomeSearch } from "./HomeSearch";
