+++ packages/app/src/modules/layout/index.ts
@@ -1,5 +1,6 @@
 export { Breadcrumb } from "./Breadcrumb";
 export { Footer } from "./Footer";
 export { Header } from "./Header";
+export { ResourceBanner } from "./ResourceBanner";
 export { NewTabNotice } from "./shared/NewTabNotice";
 export { SkipLinks } from "./shared/SkipLinks";
